requirements.txt
flask==2.0.1
jinja2==3.0.1
requests==2.25.0
pyyaml==5.3
cryptography==3.4.6
pyjwt==1.7.1
pillow==8.0.0
urllib3==1.26.4
pycryptodome==3.9.0
flask-cors==3.0.10
gunicorn==20.1.0
sqlalchemy==1.4.0
